fix(handlers): catch errors sending error response back to client
If we get an error back from the proxy we try to send an error back to the client to match.  If we fail to send the error back to the client, we should simply log an error rather than abort the entire process.

See https://github.com/chimurai/http-proxy-middleware/issues/346