fix(router): securely check path for proper destination
The router previously checked whether a path contained a route key. This check
was insufficient, as it allowed a route configured as such:
    '/admin': 'http://localhost:4002',
to be accessed to be accessible by requesting `/some/public/path?q=/admin`,
which would match and route the request to localhost:4002.

If http-proxy-middleware is in front of a firewall which restricts access to
/admin, this could have allowed unauthorized access to the admin route.